backend/_requirements.txt
fastapi
uvicorn
aiofiles
pymongo
pytest
python-dateutil
setuptools
sqlalchemy
alembic
psycopg2-binary
python_multipart
python-jose[cryptograpy]
pytest
httpx
python-dotenv
pydantic_settings